+++ src/DbusContract.h
@@ -105,7 +105,6 @@ inline constexpr char interfaceName[] = "quest.entropy.PlasmaVPN.ControlCenter1"
 
 namespace Method
 {
-inline constexpr char quit[] = "Quit";
 inline constexpr char requestRunnerAction[] = "RequestRunnerAction";
 inline constexpr char showControlCenter[] = "ShowControlCenter";
 inline constexpr char showSettings[] = "ShowSettings";